Adding ETH-ERC20 tests for creditlines
