+++ contracts/yield/CompoundYield.sol
@@ -9,6 +9,7 @@ import '@openzeppelin/contracts/math/SafeMath.sol';
 import '../interfaces/IYield.sol';
 import '../interfaces/Invest/ICEther.sol';
 import '../interfaces/Invest/ICToken.sol';
+
 /**
  * @title Yield contract
  * @notice Implements the functions to lock/unlock tokens into available exchanges
